[Gold II] Title: 꿈틀꿈틀 호석 애벌레 - 효율성, Time: 160 ms, Memory: 16156 KB, Score: 66 point -BaekjoonHub

+++ 백준/Gold/20181. 꿈틀꿈틀 호석 애벌레 － 효율성/꿈틀꿈틀 호석 애벌레 － 효율성.java
@@ -6,15 +6,15 @@ public class Main {
 	public static void main(String[] args) throws IOException {
 		int N = read();
 		int K = read();
-		int[] list = new int[N + 1];
+		long[] list = new long[N + 1];
 		for (int i = 1; i <= N; i++) list[i] = list[i - 1] + read();
-		int[] dp = new int[N + 1];
+		long[] dp = new long[N + 1];
 		
 		int pointer = 0;
 		for (int i = 1; i <= N; i++) {
 			dp[i] = dp[i - 1];
 			while (list[i - 1] - list[pointer] >= K) pointer++;
-			int foods = list[i] - list[pointer] - K;
+			long foods = list[i] - list[pointer] - K;
 			if(foods >= 0) dp[i] = Math.max(dp[i], dp[pointer] + foods);
 		}
